+++ src/index.ts
@@ -4,6 +4,7 @@ export * from './firestore/index.js';
 export * from './identity-platform/index.js';
 export * from './logging/index.js';
 export * from './pubsub/index.js';
+export * from './scheduler/index.js';
 export * from './spanner/index.js';
 export * from './tasks/index.js';
 export * from './transaction/index.js';