+++ src/tasks/index.ts
@@ -1,4 +1,10 @@
+export { CloudTasksInfo } from './cloud-tasks-info.js';
 export * from './errors.js';
+export {
+  CLOUD_TASKS_EVENT_HANDLER_ID,
+  CloudTasksEventHandlerInterceptor,
+} from './interceptor.js';
 export { CloudTasksModule } from './module.js';
 export { CloudTasksScheduler, HttpMethod } from './scheduler.js';
 export type { HttpRequest, Task } from './scheduler.js';
+export { CloudTasksEventInfo } from './task-event-info.decorator.js';