+++ src/testing.spec.ts
@@ -22,15 +22,18 @@ import {
   FirestoreCollectionsModule,
 } from './firestore/index.js';
 import { PubSubPublisherModule } from './pubsub/index.js';
+import { CloudSchedulerFixture } from './scheduler/testing.js';
 import {
   SpannerColumn,
   SpannerEntityManager,
   SpannerModule,
   SpannerTable,
 } from './spanner/index.js';
+import { CloudTasksFixture } from './tasks/testing.js';
 import {
   AuthUsersFixture,
   createGoogleFixtures,
+  FirebaseFixture,
   FirestoreFixture,
   PubSubFixture,
 } from './testing.js';
@@ -258,6 +261,14 @@ describe('AppFixture', () => {
       });
 
       expect(() => fixture.get(VersionedEntityFixture)).toThrow();
+    });
+
+    it('should provide additional fixtures', () => {
+      expect(fixture.get(FirebaseFixture)).toBeInstanceOf(FirebaseFixture);
+      expect(fixture.get(CloudTasksFixture)).toBeInstanceOf(CloudTasksFixture);
+      expect(fixture.get(CloudSchedulerFixture)).toBeInstanceOf(
+        CloudSchedulerFixture,
+      );
     });
   });