+++ src/testing.ts
@@ -4,7 +4,9 @@ import type { Type } from '@nestjs/common';
 import { FirestoreFixture } from './firestore/testing.js';
 import { AuthUsersFixture } from './identity-platform/testing.js';
 import { PubSubFixture } from './pubsub/testing.js';
+import { CloudSchedulerFixture } from './scheduler/testing.js';
 import { SpannerFixture } from './spanner/testing.js';
+import { CloudTasksFixture } from './tasks/testing.js';
 import { AppCheckFixture, FirebaseFixture } from './testing.js';
 import {
   FirestorePubSubTransactionRunner,
@@ -79,5 +81,7 @@ export function createGoogleFixtures(
     new PubSubFixture(options.pubSubTopics ?? {}),
     ...(disableAppCheck ? [new AppCheckFixture()] : []),
     ...versionedEntityFixture,
+    new CloudTasksFixture(),
+    new CloudSchedulerFixture(),
   ];
 }